requirements.txt
alpaca-py
pandas
python-dotenv
pytest
backtesting
